+++ CHANGELOG.md
@@ -2,32 +2,78 @@
 
 All notable changes to this project will be documented in this file.
 
-## [2.10.0-pre-68.3](https://github.com/SSmale/Duux-Home-Assistant/compare/v2.10.0-pre-68.2...v2.10.0-pre-68.3) (2026-06-12)
+## [2.11.0](https://github.com/SSmale/Duux-Home-Assistant/compare/v2.10.0...v2.11.0) (2026-06-12)
 
 ### Added
 
-* gets the ultimate fan setup ([0715f28](https://github.com/SSmale/Duux-Home-Assistant/commit/0715f28b7ff859ca1e68348a746290378f9dfa46))
+* Add Duux Neo Humidifier support ([#55](https://github.com/SSmale/Duux-Home-Assistant/issues/55)) ([#58](https://github.com/SSmale/Duux-Home-Assistant/issues/58)) ([9562173](https://github.com/SSmale/Duux-Home-Assistant/commit/95621732a4505c65aa42493ebc073f05d59044ce))
+* adds the Whisper Fans ([#51](https://github.com/SSmale/Duux-Home-Assistant/issues/51)) ([78395c8](https://github.com/SSmale/Duux-Home-Assistant/commit/78395c8ec493975446fd8111403ed26e7d1eb885)), closes [#69](https://github.com/SSmale/Duux-Home-Assistant/issues/69) [#70](https://github.com/SSmale/Duux-Home-Assistant/issues/70)
 
-## [2.10.0-pre-68.2](https://github.com/SSmale/Duux-Home-Assistant/compare/v2.10.0-pre-68.1...v2.10.0-pre-68.2) (2026-06-12)
+## [2.11.0-pre-55.3](https://github.com/SSmale/Duux-Home-Assistant/compare/v2.11.0-pre-55.2...v2.11.0-pre-55.3) (2026-06-12)
 
 ### Fixed
 
-* fixes merge issue ([8e777c2](https://github.com/SSmale/Duux-Home-Assistant/commit/8e777c23decd1d0381b834f210e308e1f9306b78))
+* fixes merge issue with import ([3e5d4e7](https://github.com/SSmale/Duux-Home-Assistant/commit/3e5d4e78c9e1754574e080adbf07b90962ff9190))
 
-## [2.10.0-pre-68.1](https://github.com/SSmale/Duux-Home-Assistant/compare/v2.9.0...v2.10.0-pre-68.1) (2026-06-12)
+## [2.11.0-pre-55.2](https://github.com/SSmale/Duux-Home-Assistant/compare/v2.11.0-pre-55.1...v2.11.0-pre-55.2) (2026-06-12)
+
+### Fixed
+
+* fixes manifest ([5cb9644](https://github.com/SSmale/Duux-Home-Assistant/commit/5cb9644f925015539a31d7465094248bb6a2ff3f))
+
+## [2.11.0-pre-55.1](https://github.com/SSmale/Duux-Home-Assistant/compare/v2.10.0...v2.11.0-pre-55.1) (2026-06-12)
 
 ### Added
 
-* adds flex ultimate ([70dc23b](https://github.com/SSmale/Duux-Home-Assistant/commit/70dc23b3979824c43f9868ad582f24f9f311ba36))
+* Add Duux Neo Humidifier support ([#55](https://github.com/SSmale/Duux-Home-Assistant/issues/55)) ([e65fd09](https://github.com/SSmale/Duux-Home-Assistant/commit/e65fd09b6237bca331c103080906423f6292c6b0))
+* adds neo support ([332769f](https://github.com/SSmale/Duux-Home-Assistant/commit/332769fa003072808539898b4542bc582f7901f3))
+* adds the Whisper Fans ([#51](https://github.com/SSmale/Duux-Home-Assistant/issues/51)) ([78395c8](https://github.com/SSmale/Duux-Home-Assistant/commit/78395c8ec493975446fd8111403ed26e7d1eb885)), closes [#69](https://github.com/SSmale/Duux-Home-Assistant/issues/69) [#70](https://github.com/SSmale/Duux-Home-Assistant/issues/70)
+
+## [2.11.0-pre-50.2](https://github.com/SSmale/Duux-Home-Assistant/compare/v2.11.0-pre-50.1...v2.11.0-pre-50.2) (2026-06-12)
+
+### Fixed
+
+* reverts merge overrite ([23c9836](https://github.com/SSmale/Duux-Home-Assistant/commit/23c9836555dcfb0976f46441fbef0ff51c626412))
+
+## [2.11.0-pre-50.1](https://github.com/SSmale/Duux-Home-Assistant/compare/v2.10.0...v2.11.0-pre-50.1) (2026-06-12)
+
+### Added
+
 * adds initial support for elevate ([#70](https://github.com/SSmale/Duux-Home-Assistant/issues/70)) ([cc97983](https://github.com/SSmale/Duux-Home-Assistant/commit/cc979835eebf32e34f39d283d88182a79423e6cf))
 * adds ossolation for fans ([8b5dc49](https://github.com/SSmale/Duux-Home-Assistant/commit/8b5dc49340123a03ca0c2c025ca87005e2bd9a57))
 * adds support for the whisper fans ([8e7bcb2](https://github.com/SSmale/Duux-Home-Assistant/commit/8e7bcb29a565d62e5de496325c4384d2dd2a75b6))
 
 ### Fixed
 
-* adds explicit features for the fan settings ([fb4c03b](https://github.com/SSmale/Duux-Home-Assistant/commit/fb4c03b733c137176f7671f82437723b94bee646))
 * adds explicit features for the fan settings ([#69](https://github.com/SSmale/Duux-Home-Assistant/issues/69)) ([639c28c](https://github.com/SSmale/Duux-Home-Assistant/commit/639c28c335aaa25e159b9fe71acf32a63a618848))
+* fixes merge issue ([8e777c2](https://github.com/SSmale/Duux-Home-Assistant/commit/8e777c23decd1d0381b834f210e308e1f9306b78))
+* merge issues ([fa31b83](https://github.com/SSmale/Duux-Home-Assistant/commit/fa31b83f5452e344606ca0e7af613309db0418bd))
+
+## [2.10.0](https://github.com/SSmale/Duux-Home-Assistant/compare/v2.9.0...v2.10.0) (2026-06-12)
+
+### Added
+
+* adds support for the beam mini ([#37](https://github.com/SSmale/Duux-Home-Assistant/issues/37)) ([1dcaa20](https://github.com/SSmale/Duux-Home-Assistant/commit/1dcaa2029daeb41b12ec813e0ca8c0fcaa38c305))
+* release trigger ([9e6718e](https://github.com/SSmale/Duux-Home-Assistant/commit/9e6718e4d905986011cb85721cabcdf5033a3e69))
+
+### Fixed
+
 * **climate, humidifier:** mark entity unavailable when coordinator data is None ([#63](https://github.com/SSmale/Duux-Home-Assistant/issues/63)) ([#65](https://github.com/SSmale/Duux-Home-Assistant/issues/65)) ([#67](https://github.com/SSmale/Duux-Home-Assistant/issues/67)) ([2bd3907](https://github.com/SSmale/Duux-Home-Assistant/commit/2bd39071f9ffe8a4b36b4be063b228d4d1b3d360))
+
+## [2.10.0-alpha.1](https://github.com/SSmale/Duux-Home-Assistant/compare/v2.9.0...v2.10.0-alpha.1) (2026-06-12)
+
+### Added
+
+* adds support for the beam mini ([0d423fb](https://github.com/SSmale/Duux-Home-Assistant/commit/0d423fb6d03c57e6bd9c93055d0cca901397e4e7))
+* code issue tidy up ([000cea9](https://github.com/SSmale/Duux-Home-Assistant/commit/000cea9058726ec77a334466c2e877f3bd1a6564))
+* more features of the Beam implemented ([83fcab8](https://github.com/SSmale/Duux-Home-Assistant/commit/83fcab84256a4bfbaa7882416be1e5943ee87498))
+
+### Fixed
+
+* adds a new base class for dehumidifier and humidifier ([570a4c6](https://github.com/SSmale/Duux-Home-Assistant/commit/570a4c649d6ef062f319a68d10d0c46f64ab3764))
+* **climate, humidifier:** mark entity unavailable when coordinator data is None ([#63](https://github.com/SSmale/Duux-Home-Assistant/issues/63)) ([#65](https://github.com/SSmale/Duux-Home-Assistant/issues/65)) ([#67](https://github.com/SSmale/Duux-Home-Assistant/issues/67)) ([2bd3907](https://github.com/SSmale/Duux-Home-Assistant/commit/2bd39071f9ffe8a4b36b4be063b228d4d1b3d360))
+* the import was needed for the Bora ([60ca64d](https://github.com/SSmale/Duux-Home-Assistant/commit/60ca64dfb25f0c1ab022ea16e27cf862c0223536))
+* used the HA import for the mode ([a974655](https://github.com/SSmale/Duux-Home-Assistant/commit/a974655f36a3e62b2158b9a0f9691fac2e812f66))
 
 ## [2.10.0-pre-44.3](https://github.com/SSmale/Duux-Home-Assistant/compare/v2.10.0-pre-44.2...v2.10.0-pre-44.3) (2026-06-12)
 
@@ -146,5 +192,3 @@ All notable changes to this project will be documented in this file.
 * [ ] Support for Duux fans
 * [ ] Local API support (if available)
 * [ ] Energy dashboard integration
-
-